requirements.txt
openai
pyaudio
keyboard
pyperclip
requests
python-dotenv
Pillow
customtkinter